Extract entries.fetchFullContent business logic into the full-content service (#956)
The fetch/hash/sanitize/persist/narration-invalidation flow (~140 lines)
lived inline in the entries router with only the raw HTTP fetch in a
service. Move it to fetchAndStoreFullContent in services/full-content.ts
so it follows the services-layer pattern and is reusable outside tRPC.

The write to the shared entries row (including fullContentError) is kept
and now documented as deliberate: full content and its fetchability are
properties of the source URL, not the requesting user, and sharing the
result avoids duplicate fetches by other subscribers.
